Revert "Add "//apex_available:platform" to apex_available."
This reverts commit e3de1d1b704422e792bd385e2c45d8450bb42bae.

It also reverts "Add some missing apexes to apex_available"
commit 16849093aab0f7796097061cd78eeac937acf795.

They are being reverted together because the first change caused some
build issues and combining them eliminates any risk of them causing
other build issues by not being applied atomically.

Reason for revert: Not needed

The original change was made in an attempt to close a perceived hole in
the apex availability checking. However, there is no such hole and so
this change is unnecessary. While there is a dependency path from the
apex to this module one of the dependencies in that path is from a
static library to a shared library. That dependency does not result in
the code that includes the static library being linked to the shared
library at either build time or runtime. Therefore, the shared library
is not part of the APEX and so does not need to be marked as such.

Bug: 152762638
Test: m droid
Change-Id: Ib8218f2fbc26e2b0fd01dbc0668bfdfe359aaefa
Exempt-From-Owner-Approval: revert